requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.0.1
numpy==1.24.3
transformers==4.35.2
torch==2.1.0
eas# Core FastAPI dependencies
fastapi
uvicorn
python-multipart

# Image processing
opencv-python
pillow
numpy

# OCR libraries
transformers
torch
torchvision
easyocr

# Optional: Only if you need JWT authentication later
# python-jose[cryptography]yocr==1.7.0
python-jose[cryptography]==3.3.0